requirements.txt
anyio==4.3.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
filelock==3.13.1
Flask==3.0.2
fsspec==2024.2.0
h11==0.14.0
hkdf==0.0.3
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.3
llvmlite==0.42.0
magic-wormhole==0.13.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
ollama==0.2.0
openai-whisper==20231117
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==24.1.0
regex==2023.12.25
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
spake2==0.8
sympy==1.12
tiktoken==0.6.0
torch==2.2.0
tqdm==4.66.2
triton==2.2.0
Twisted==24.3.0
txaio==23.1.1
txtorcon==23.11.0
typing_extensions==4.9.0
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.1
zope.interface==6.2
